Vector Field Plotting Code
This class takes in co-ordinate systems and stores them in spherical or cartesian form.  The class is intended to be a parent class for a specific vector field class that takes in the co-ordinates and spits out a vector field.  It'll probably make more sense to change the name to "Coordinates" and then name the actual vector field children accordingly